feat: adding current code report analysis and linting changes
