+++ go/goutils/commonutils/sort.go
@@ -27,13 +27,14 @@ func (s asInt) Less(i, j int) bool {
 }
 
 func SortKeysAsNumber(unsortedMap *map[string]string) *[]string {
-
-	//Sort DAGSegments by their height and then process.
+	// Sort DAGSegments by their height and then process.
 	sortedKeys := make([]string, 0, len(*unsortedMap))
 	for k := range *unsortedMap {
 		sortedKeys = append(sortedKeys, k)
-		//log.Debugf("Key %s", k)
+		// log.Debugf("Key %s", k)
 	}
+
 	sort.Sort(asInt(sortedKeys))
+
 	return &sortedKeys
 }
